server/lib/lib.pro: Rename library and remove dependency on main/ folder.
It makes more sens to call the library 'libopenjabnab' instead of libcommon,
so rename it. The lead to a bunch of modification of the other projects files.

While looking at it, noticed that the library was needing the main application
folder in order to include 'openjabnab.h' which is surprising. Turns out that
it was used to include other headers. So, modify 'openjabnab.h' users to use
the right headers, and remove now useless headers include from openjabnab.h.

Signed-off-by: Arnaud Patard <arnaud.patard@rtp-net.org>