server/lib: Allow to use more standards paths
The current code is using only the directory containing the
openjabnab binary to store and find files like plugins or settings
file. It's working fine the server/bin directory is used as-is,
but if someone wants more standards paths like /usr/bin/openjabnab,
things well break.

This commit is introducing helpers to find the needed directories,
eg QStandardsPaths for settings and QLibraryInfo::LibrariesPath for
the plugins. In order to not break things, some (fragile) code has
been added to try detecting the current behaviour and keep it working.

Signed-off-by: Arnaud Patard <arnaud.patard@rtp-net.org>